refactor(workflow): extract implementation state codecs
